added boxdecoration code
